requirements.txt
playwright
supabase
python-dotenv
